// dekko-client/src/app/user/model/user.model.ts
export class UserModel {

    id: number|undefined;
    firstName: string|undefined;
    lastName: string|undefined;
    email: string|undefined;
    mobileNumber: string|undefined;
    address: string|undefined;
    dateOfBirth: Date|undefined;
    userId: string|undefined;
    password: string|undefined;
    
}